Handle short posts in pager.
This commit really does two things.

First, it reworks how the pager determines the post for the page by
consulting the previous pager instance, and then counting up from there,
rather than calculating indexes.  This removes the requirement that
pages all have the same number of posts.

Second, it adds a weighting function to pages and uses that to determine
the number of posts for a given page and the number of total posts.

The idea is that a page may be composed of a large number of short posts
and a few long posts, with the number of each determined by a weight
that ultimately acts as a ratio.