Support showing hidden posts selectively.
This change adds a 'show_hidden' setting that can be used to display
hidden posts in paginated pages.

This setting is by default *enabled for autopages*.  The thinking, here,
is that for category pages or whatnot, you want to see everything.
Obviously not everyone will like that default, but this is my fork,
so...